+++ packages/sdks/typescript/CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `messages` parameter to `prompts.run()` method. This allows appending any kind of messages to the conversation after the compiled prompt. Note: This is not compatible with the `<step>` feature of PromptL.
+
 ### Deprecated
 
-- The `userMessage` parameter in `prompts.run()` is now deprecated. Use the `messages` parameter instead (coming soon). The `userMessage` parameter will be removed in a future version.
+- The `userMessage` parameter in `prompts.run()` is now deprecated. Use the `messages` parameter instead. The `userMessage` parameter will be removed in a future version.
 
 ## [5.4.1] - 2026-01-30